requirements.txt
anthropic>=0.50.0
canvasapi>=3.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
playwright>=1.40.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0
